raspberry/afssDocker/requirements.txt
pyserial
datetime
requests
